multimodal_ai/main.py
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

client = OpenAI()

response = client.responses.create(
    model="gpt-4.1-mini",
    input=[{
        "role": "user",
        "content": [
            {"type": "input_text", "text": "what's in this image?"},
            {
                "type": "input_image",
                "image_url": "https://images.pexels.com/photos/3987011/pexels-photo-3987011.jpeg",
            },
        ],
    }],
)

print(response.output_text)